% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandCap.R
\name{RandCapProd}
\alias{RandCapProd}
\title{Generate Production Randomization}
\usage{
RandCapProd(randomization_object, seed)
}
\arguments{
\item{randomization_object}{A list containing randomization settings and tables, created by \code{RandCapGen}.}

\item{seed}{An integer specifying the seed for reproducibility.}
}
\value{
A modified \code{randomization_object} with updated tables and settings for production.
}
\description{
This function finalizes the production randomization based on the development settings
in the \code{randomization_object}. It shuffles the blocks and generates production-ready randomization tables.
}
\examples{
# Assuming you have a randomization object from RandCapGen
RandCapProd(randomization_object = my_randomization_object, seed = 123)
}
\seealso{
\code{\link{RandCapGen}},
\code{\link{RandCapSettings}},
\code{\link{RandCapTable}},
and \code{\link{RandCapBalance}}
}
