% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandCap.R
\name{RandCapTable}
\alias{RandCapTable}
\title{Export Randomization Tables for REDCap}
\usage{
RandCapTable(
  randomization_object,
  save_for_REDCap = FALSE,
  save_random_table = FALSE
)
}
\arguments{
\item{randomization_object}{A list. The randomization object generated by \code{RandCapGen}.}

\item{save_for_REDCap}{Logical. If \code{TRUE}, saves the simplified table for REDCap as a CSV file.}

\item{save_random_table}{Logical. If \code{TRUE}, saves the complete randomization table as a CSV file.}
}
\value{
A list containing two data frames:
\describe{
\item{simplified_table}{A simplified dataset ready for REDCap import.}
\item{full_table}{The complete randomization dataset.}
}
}
\description{
\code{RandCapTable} saves the randomization tables from a \code{RandCapGen} object to CSV files, one for a complete randomization list and another simplified for REDCap import.
}
\examples{
# Example usage to save both tables
RandCapTable(randomization_object = my_randomization_object, save_for_REDCap = TRUE, save_random_table = TRUE)
}
\seealso{
\code{\link{RandCapGen}},
\code{\link{RandCapBalance}},
\code{\link{RandCapSettings}},
and \code{\link{RandCapProd}}
}
